[Fix] 로그인 페이지 배포 테스트
