--- src/components/domain/signin/SigninButton/SigninButton.module.scss
@@ -1,27 +1,0 @@
-.btn {
-  padding: 1.6rem 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 2rem;
-
-  border-radius: 1rem;
-  width: 100%;
-  font-size: 1.6rem;
-  font-weight: 500;
-  letter-spacing: -0.00375rem;
-
-  &.kakao {
-    background-color: $kakao-bg;
-    color: $kakao-text;
-  }
-  &.naver {
-    background-color: $naver-bg;
-    color: $white;
-  }
-  &.google {
-    background-color: $white;
-    color: $google-text;
-    border: 1px solid $google-boder;
-  }
-}
