--- src/lib/apis/axios.tsx
@@ -1,5 +1,0 @@
-import axios from "axios";
-
-export const instance = axios.create({
-  baseURL: "https://api.bin-finder.net",
-});
