+++ src/lib/atoms/userAtom.ts
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+
+//추후 수정
+export const userAtom = atom<any>(null);
